requirements.txt
# Speech-To-Text solution
openai-whisper==20240930

# Four sounds
PyAudio==0.2.14

# For the console client to use microphone
SpeechRecognition==3.12.0

soundfile==0.12.1   # to read audio wave streamed
sounddevice==0.5.1 # For the console client to play audio
numpy==1.26.4       # to convert audio data

# LLM
ollama==0.4.4
asyncio==3.4.3

# For torch with CUDA, depends on GPU and driver !! Check cuda version to install
# pip3 install torch torchaudio --index-url https://download.pytorch.org/whl/cu124
# pip install torch --index-url https://download.pytorch.org/whl/cu124

# Logging
# logging installed by others

# Text-To-Speech solution
coqui-tts
# kokoro

# API web server
fastapi
uvicorn
python-multipart
websockets

#Telegram API
pyTelegramBotAPI
ffmpeg-python
